Refactored rest endpoints to respond to correct verbs + http statuses

+++ src/main/java/paralaks_gmail_com/todolist/controller/TodoListController.java
@@ -1,11 +1,13 @@
 package paralaks_gmail_com.todolist.controller;
 
+import org.springframework.http.ResponseEntity;
 import paralaks_gmail_com.todolist.model.TodoList;
 import paralaks_gmail_com.todolist.service.TodoListService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/todolists")
@@ -17,14 +19,15 @@ public class TodoListController {
     todoListService = service;
   }
 
-  @RequestMapping("/")
-  public List<TodoList> list() {
-    return todoListService.list();
+  @GetMapping("/")
+  public ResponseEntity<List<TodoList>> list() {
+    return ResponseEntity.ok(todoListService.list());
   }
 
-  @RequestMapping("/{id}")
-  public TodoList findOne(@PathVariable Integer id) {
-    return todoListService.findOne(id).orElse(null);
+  @GetMapping("/{id}")
+  public ResponseEntity<TodoList> findOne(@PathVariable Integer id) {
+    Optional<TodoList> list = todoListService.findOne(id);
+    return list.isPresent() ? ResponseEntity.ok(list.get()) : ResponseEntity.notFound().build();
   }
 
   @PostMapping("/add")
@@ -33,7 +36,12 @@ public class TodoListController {
   }
 
   @DeleteMapping("/{id}")
-  public String deleteOne(@PathVariable Integer id) {
-    return todoListService.deleteOne(id);
+  public ResponseEntity<String> deleteOne(@PathVariable Integer id) {
+    if (id == null)
+      return ResponseEntity.notFound().build();
+
+    String result = todoListService.deleteOne(id);
+
+    return result == null ? ResponseEntity.notFound().build() : ResponseEntity.ok(result);
   }
 }
